pyproject.toml
[tool.poetry]
name = "err-discord-bridge"
version = "0.1.0"
description = "Errbot plugin to bridge your backend to discord"
authors = ["Andrew Herrington <andrew.the.techie@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "discord_bridge"}]
include = ["*.plug"]

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.3.2"
pydantic = "^2.7.1"
toml = "^0.10.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
coverage = "^7.4.4"
pytest-cov = "^5.0.0"
pre-commit = "^3.7.0"
ruff = "^0.3.7"
pytest-xdist = "^3.5.0"
bandit = "^1.7.8"
pyupgrade = "^3.15.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.plugins."errbot.plugins"]
discord_bridge = "discord_bridge:DiscordBridge"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.pytest.ini_options]
norecursedirs = ".github ci .git .idea"
addopts = "-n auto --cov=discord_bridge --cov-report xml:.coverage.xml --cov-report=term-missing"

[tool.bandit.assert_used]
skips = ['*/test_*.py']
